requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
azure-storage-blob>=12.16.0
azure-core~=1.34.0
autogluon